+++ packages/driver/src/cy/commands/actions/type.ts
@@ -31,7 +31,6 @@ export default function (Commands, Cypress, cy, state, config) {
       parseSpecialCharSequences: true,
       waitForAnimations: config('waitForAnimations'),
       animationDistanceThreshold: config('animationDistanceThreshold'),
-      scrollBehavior: config('scrollBehavior'),
     })
 
     if (options.log) {
@@ -273,6 +272,11 @@ export default function (Commands, Cypress, cy, state, config) {
       const isContentEditable = $elements.isContentEditable(options.$el.get(0))
       const isTextarea = $elements.isTextarea(options.$el.get(0))
 
+      // click event is only fired on button, image, submit, reset elements.
+      // That's why we cannot use $elements.isButtonLike() here.
+      const type = (type) => $elements.isInputType(options.$el.get(0), type)
+      const sendClickEvent = type('button') || type('image') || type('submit') || type('reset')
+
       return keyboard.type({
         $el: options.$el,
         chars,
@@ -348,13 +352,24 @@ export default function (Commands, Cypress, cy, state, config) {
           })
         },
 
-        onEnterPressed (id) {
+        onEnterPressed (el) {
           // dont dispatch change events or handle
           // submit event if we've pressed enter into
           // a textarea or contenteditable
-
           if (isTextarea || isContentEditable) {
             return
+          }
+
+          // https://github.com/cypress-io/cypress/issues/19541
+          // Send click event on type('{enter}')
+          if (sendClickEvent) {
+            // Firefox sends a click event automatically.
+            if (!Cypress.isBrowser('firefox')) {
+              const ctor = $dom.getDocumentFromElement(el).defaultView?.PointerEvent
+              const event = new ctor('click')
+
+              el.dispatchEvent(event)
+            }
           }
 
           // if our value has changed since our
@@ -363,7 +378,7 @@ export default function (Commands, Cypress, cy, state, config) {
           const changeEvent = state('changeEvent')
 
           if (changeEvent) {
-            changeEvent(id)
+            changeEvent()
           }
 
           // handle submit event handler here
@@ -415,7 +430,7 @@ export default function (Commands, Cypress, cy, state, config) {
         }
       }
 
-      return $actionability.verify(cy, options.$el, options, {
+      return $actionability.verify(cy, options.$el, config, options, {
         onScroll ($el, type) {
           return Cypress.action('cy:scrolled', $el, type)
         },
@@ -442,6 +457,7 @@ export default function (Commands, Cypress, cy, state, config) {
             timeout: options.timeout,
             interval: options.interval,
             errorOnSelect: false,
+            scrollBehavior: options.scrollBehavior,
           })
           .then(() => {
             let activeElement = $elements.getActiveElByDocument($elToClick)
@@ -502,7 +518,6 @@ export default function (Commands, Cypress, cy, state, config) {
       force: false,
       waitForAnimations: config('waitForAnimations'),
       animationDistanceThreshold: config('animationDistanceThreshold'),
-      scrollBehavior: config('scrollBehavior'),
     })
 
     // blow up if any member of the subject
@@ -569,7 +584,7 @@ export default function (Commands, Cypress, cy, state, config) {
           notReadonly: true,
         }
 
-        return $actionability.verify(cy, $el, options, {
+        return $actionability.verify(cy, $el, config, options, {
           onScroll ($el, type) {
             return Cypress.action('cy:scrolled', $el, type)
           },