+++ packages/driver/src/cy/commands/fixtures.ts
@@ -2,10 +2,15 @@
 
 import _ from 'lodash'
 import Promise from 'bluebird'
+import { basename } from 'path'
 
 import $errUtils from '../../cypress/error_utils'
 
 const clone = (obj) => {
+  if (Buffer.isBuffer(obj)) {
+    return Buffer.from(obj)
+  }
+
   return JSON.parse(JSON.stringify(obj))
 }
 
@@ -47,7 +52,7 @@ export default (Commands, Cypress, cy, state, config) => {
         options = args[1]
       }
 
-      if (_.isString(args[0])) {
+      if (_.isString(args[0]) || args[0] === null) {
         options.encoding = args[0]
       }
 
@@ -64,9 +69,24 @@ export default (Commands, Cypress, cy, state, config) => {
           return $errUtils.throwErr(response.__error)
         }
 
+        // https://github.com/cypress-io/cypress/issues/1558
+        // We invoke Buffer.from() in order to transform this from an ArrayBuffer -
+        // which socket.io uses to transfer the file over the websocket - into a
+        // `Buffer`, which webpack polyfills in the browser.
+        if (options.encoding === null) {
+          response = Buffer.from(response)
+        } else if (response instanceof ArrayBuffer) {
+          // Cypress' behavior is to base64 encode binary files if the user
+          // doesn't explicitly pass `null` as the encoding.
+          response = Buffer.from(response).toString('base64')
+        }
+
         // add the fixture to the cache
         // so it can just be returned next time
         cache[fixture] = response
+
+        // Add the filename as a symbol, in case we need it later (such as when storing an alias)
+        state('current').set('fileName', basename(fixture))
 
         // return the cloned response
         return clone(response)