+++ packages/server/lib/browsers/chrome.ts
@@ -207,7 +207,7 @@ const _normalizeArgExtensions = function (extPath, args, pluginExtensions, brows
     userExtensions = userExtensions.concat(pluginExtensions)
   }
 
-  const extensions = [].concat(userExtensions, extPath, pathToTheme)
+  const extensions = ([] as any).concat(userExtensions, extPath, pathToTheme)
 
   args.push(LOAD_EXTENSION + _.compact(extensions).join(','))
 
@@ -520,13 +520,13 @@ export = {
     const originalBrowserKill = launchedBrowser.kill
 
     /* @ts-expect-error */
-    launchedBrowser.kill = async (...args) => {
+    launchedBrowser.kill = (...args) => {
       debug('closing remote interface client')
 
-      await criClient.close()
+      criClient.close()
       debug('closing chrome')
 
-      await originalBrowserKill.apply(launchedBrowser, args)
+      originalBrowserKill.apply(launchedBrowser, args)
     }
 
     await this._maybeRecordVideo(criClient, options, browser.majorVersion)