+++ packages/server/lib/modes/run.js
@@ -608,8 +608,21 @@ const openProjectCreate = (projectRoot, socketId, args) => {
   return openProject.create(projectRoot, args, options)
 }
 
-const createAndOpenProject = async function (socketId, options) {
+async function checkAccess (folderPath) {
+  return fs.access(folderPath, fs.W_OK).catch((err) => {
+    if (['EACCES', 'EPERM'].includes(err.code)) {
+      // we cannot write due to folder permissions
+      return errors.warning('FOLDER_NOT_WRITABLE', folderPath)
+    }
+
+    throw err
+  })
+}
+
+const createAndOpenProject = async (socketId, options) => {
   const { projectRoot, projectId } = options
+
+  await checkAccess(projectRoot)
 
   return openProjectCreate(projectRoot, socketId, options)
   .then((open_project) => open_project.getProject())
@@ -1415,7 +1428,6 @@ module.exports = {
     })
 
     if (browser.family !== 'chromium' && !options.config.chromeWebSecurity) {
-      console.log()
       errors.warning('CHROME_WEB_SECURITY_NOT_SUPPORTED', browser.family)
     }
 
@@ -1447,7 +1459,7 @@ module.exports = {
           compressedVideoName: videoRecordProps.compressedVideoName,
           endVideoCapture: videoRecordProps.endVideoCapture,
           startedVideoCapture: videoRecordProps.startedVideoCapture,
-          exit: options.exit,
+          exit: config.exit,
           videoCompression: options.videoCompression,
           videoUploadOnPasses: options.videoUploadOnPasses,
           quiet: options.quiet,
@@ -1595,7 +1607,6 @@ module.exports = {
               video: config.video,
               videoCompression: config.videoCompression,
               videoUploadOnPasses: config.videoUploadOnPasses,
-              exit: options.exit,
               headed: options.headed,
               quiet: options.quiet,
               outputPath: options.outputPath,